+++ test/script/lint.sh
@@ -1,0 +1,12 @@
+#!/bin/sh
+
+. ./config.sh
+
+if [ "${ORION}" != "orion" ]; then
+  echo "Error ORION value in config.sh"
+fi
+
+if [ "${KEYROCK}" != "keyrock" ]; then
+  echo "Error ORION value in config.sh"
+fi
+