+++ test/script/lint-yamllint.sh
@@ -31,17 +31,4 @@ set -ue
 cd "$(dirname "$0")"
 cd ../..
 
-echo "*** lint config.sh ***"
-./test/script/lint-config.sh
-
-echo "*** lint dockerfile ***"
-./test/script/lint-dockerfile.sh
-
-echo "*** lint documentation ***"
-./test/script/lint-docs.sh
-
-echo "*** lint shellcheck ***"
-./test/script/lint-shellcheck.sh
-
-echo "*** lint yamllint ***"
-./test/script/lint-yamllint.sh
+yamllint setup/templeate/docker/