+++ lets-fiware.sh
@@ -187,6 +187,7 @@ CERTBOT_DIR=${CERTBOT_DIR}
 CONFIG_DIR=${CONFIG_DIR}
 NGINX_SITES=${NGINX_SITES}
 SETUP_DIR=${SETUP_DIR}
+WORK_DIR=${WORK_DIR}
 TEMPLEATE=${TEMPLEATE}
 
 LOG_DIR=${LOG_DIR}
@@ -651,8 +652,15 @@ setup_test_cert() {
 wait() {
   logging_info "${FUNCNAME[0]}"
 
-  echo "Wait for $1"
-  while [ "404" != "$(curl http://"$1"/ -o /dev/null -w '%{http_code}\n' -s)" ]
+  local host
+  local ret
+
+  host=$1
+  shift
+  ret=$1
+
+  echo "Wait for ${host} to be ready"
+  while [ "${ret}" != "$(curl "${host}" -o /dev/null -w '%{http_code}\n' -s)" ]
   do
     sleep 1
   done
@@ -671,7 +679,7 @@ get_cert() {
   fi
 
   if sudo [ ! -d "${CERT_DIR}/live/$1" ]; then
-    wait "$1"
+    wait "http://$1/" "404"
     # shellcheck disable=SC2086
     sudo docker run --rm -v "${CERTBOT_DIR}/$1:/var/www/html/$1" -v "${CERT_DIR}:/etc/letsencrypt" "${CERTBOT}" certonly ${CERT_TEST} --agree-tos -m "${CERT_EMAIL}" --webroot -w "/var/www/html/$1" -d "$1"
   else
@@ -861,10 +869,7 @@ EOF
 
   sudo "${DOCKER_COMPOSE}" -f docker-keyrock.yml up -d
 
-  while [ "200" != "$(curl http://localhost:3000/ -o /dev/null -w '%{http_code}\n' -s)" ]
-  do
-    sleep 1
-  done
+  wait "http://localhost:3000/" "200"
 
   ${NGSI_GO} server add --host "${IDM}" --serverType keyrock --serverHost http://localhost:3000 --idmType idm --username "${IDM_ADMIN_EMAIL}" --password "${IDM_ADMIN_PASS}"
 }
@@ -1227,6 +1232,11 @@ boot_up_containers() {
 
   logging_info "docker-compose up -d --build"
   sudo "${DOCKER_COMPOSE}" up -d --build
+
+  if "$FIBB_TEST"; then
+    return
+  fi
+  wait "https://${KEYROCK}/" "200"
 }
 
 #