+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+## FIWARE Big Bang v0.2.0-next
+
 ## FIWARE Big Bang v0.2.0 - 20 Sep, 2021
 
 -   Improve docker install on Ubuntu (#44)